==> Avaluació_MIPredictor/01_Preprocessa_dades.Rmd <==
---
title: "01_Preprocessa_dades"
author: "Erik Andrés Reig"
date: "9/5/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Carreguem llibreries necessaries

```{r libraries, warning=F}
if (!requireNamespace("BiocManager", quietly = TRUE))
        install.packages("BiocManager")
#BiocManager::install("MMUPHin")
library(MMUPHin)
library(magrittr)
library(dplyr)
library(limma)
library(edgeR)
library(knitr)
library(kableExtra)
library(ggplot2)
library(vegan)
library(splitTools)
library(stringr)
library(caret)
library(data.table)
```

Lectura d'arxius

```{r read}
### Matrius de taxonomia ###
taxoS_raw <- read.delim("input/merged_bracken_S.txt", row.names = 1, 
                     check.names = F)
taxoS1_raw <- read.delim("input/merged_bracken_S1.txt", row.names = 1, 
                     check.names = F)

taxoS <- select(taxoS_raw,contains("num")) #només counts, no frac
taxoS1 <- select(taxoS1_raw,contains("num")) #només counts, no frac

nms <- as.data.frame(strsplit(names(taxoS), "_", fixed=T))[1,]
names(taxoS) <- nms

nms <- as.data.frame(strsplit(names(taxoS1), "_", fixed=T))[1,]
names(taxoS1) <- nms

### METADATA ###
metadata <- read.csv("input/metadades_full.csv")
# afegim variable study per avaluar i corregir efecte batch
metadata$study <- substr(metadata$sample,1,6)
# convertim categòriques a factors
metadata$region <- factor(metadata$region, levels=c("Xina", "EUA"))
metadata$seq_tool <- factor(metadata$seq_tool, levels=c("NovaSeq", "HiSeq"))
metadata$condition <- factor(metadata$condition, levels=c("Control", "CRC"))
metadata$study <- factor(metadata$study)
# reordenem per alinear mostres amb les columnes de counts
metadata$sample2 <- metadata$sample
metadata <- left_join(data.frame(sample2 = names(taxoS)), metadata)
rownames(metadata) <- metadata$sample

# ARXIU EXEMPLE KRAKEN AMB TOTES LES ESPÈCIES I SUB-ESPÈCIES DE REFSEQ

kraken <- read.delim("input/SRR15674130_report.txt")
kraken_S <- kraken[kraken$U=="S",]
kraken_S1 <- kraken[kraken$U=="S1",]
species <- str_trim(kraken_S[,length(kraken_S)])
strains <- str_trim(kraken_S1[,length(kraken_S1)])

```

Comprovem % variació explicada per batch

```{r}
D_before <- vegdist(t(taxoS))
fit_adonis_before <- adonis(D_before ~ study, data = metadata)
(taxoS_var_explained_before <- fit_adonis_before$aov.tab)
rm(D_before)
rm(fit_adonis_before)

D_before <- vegdist(t(taxoS1))
fit_adonis_before <- adonis(D_before ~ study, data = metadata)
(taxoS1_var_explained_before <- fit_adonis_before$aov.tab)
rm(D_before)
rm(fit_adonis_before)
```

L'efecte batch explica el i 26.4 i 24.7% de la variació de les dades S i S1, respectivament.

S'elimina efecte batch. No s'especificarà covariable target (condition) perquè s'utilitzaran dades per fer prediccions.

```{r }
fit_adjust_batch_S <- adjust_batch(feature_abd=taxoS, batch="study",
                                 covariates = NULL,
                                 data=metadata,
                                 control = list(verbose=TRUE))

fit_adjust_batch_S1 <- adjust_batch(feature_abd=taxoS1, batch="study",
                                 covariates = NULL,
                                 data=metadata,
                                 control = list(verbose=TRUE))

taxoS_adj <- as.data.frame(fit_adjust_batch_S$feature_abd_adj)
taxoS1_adj <- as.data.frame(fit_adjust_batch_S1$feature_abd_adj)
```

Comprovem ara % variació explicada per batch

```{r}
D_after <- vegdist(t(taxoS_adj))
fit_adonis_after <- adonis(D_after ~ study, data = metadata)
(taxoS_var_explained_after <- fit_adonis_after$aov.tab)
rm(D_after)
rm(fit_adonis_after)

D_after <- vegdist(t(taxoS1_adj))
fit_adonis_after <- adonis(D_after ~ study, data = metadata)
(taxoS1_var_explained_after <- fit_adonis_after$aov.tab)
rm(D_after)
rm(fit_adonis_after)
```

Tot i que segueix existent efecte batch en les dades (p=0.001), aquest s'ha reduit a 8.5 i 9.1%.

Normalitzem taxonomies amb MinMax i separem en train, val i test. Guardem test.

```{r}
preProcValues_S <- preProcess(taxoS_adj, method = "range")
taxoS_adj_norm <- predict(preProcValues_S, taxoS_adj)

preProcValues_S1 <- preProcess(taxoS1_adj, method = "range")
taxoS1_adj_norm <- predict(preProcValues_S1, taxoS1_adj)

index_split <- partition(metadata$condition, 
                         c(train=0.70, val=0.15, test=0.15),
                         type = "stratified",
                         split_into_list = TRUE,
                         use_names = TRUE,
                         shuffle = TRUE, seed = 42)

taxoS_adj_norm_test <- taxoS_adj_norm[,index_split$test]
taxoS1_adj_norm_test <- taxoS1_adj_norm[,index_split$test]
```

Eliminem efecte batch i normalitzem train i val per separat.

```{r}
taxoS_train <- taxoS[,index_split$train]
taxoS1_train <- taxoS1[,index_split$train]
metadata_train <- metadata[index_split$train,]
taxoS_val <- taxoS[,index_split$val]
taxoS1_val <- taxoS1[,index_split$val]
metadata_val <- metadata[index_split$val,]

fit_adjust_batch_S_train <- adjust_batch(feature_abd=taxoS_train,
                                         batch="study",
                                         covariates = NULL,
                                         data=metadata_train,
                                         control = list(verbose=TRUE))

fit_adjust_batch_S1_train <- adjust_batch(feature_abd=taxoS1_train,
                                         batch="study",
                                         covariates = NULL,
                                         data=metadata_train,
                                         control = list(verbose=FALSE))

fit_adjust_batch_S_val <- adjust_batch(feature_abd=taxoS_val,
                                         batch="study",
                                         covariates = NULL,
                                         data=metadata_val,
                                         control = list(verbose=TRUE))

fit_adjust_batch_S1_val <- adjust_batch(feature_abd=taxoS1_val,
                                         batch="study",
                                         covariates = NULL,
                                         data=metadata_val,
                                         control = list(verbose=FALSE))

taxoS_adj_train <- as.data.frame(fit_adjust_batch_S_train$feature_abd_adj)
taxoS1_adj_train <- as.data.frame(fit_adjust_batch_S1_train$feature_abd_adj)
taxoS_adj_val <- as.data.frame(fit_adjust_batch_S_val$feature_abd_adj)
taxoS1_adj_val <- as.data.frame(fit_adjust_batch_S1_val$feature_abd_adj)

preProcValues_S_train <- preProcess(taxoS_adj_train, method = "range")
preProcValues_S1_train <- preProcess(taxoS1_adj_train, method = "range")
preProcValues_S_val <- preProcess(taxoS_adj_val, method = "range")
preProcValues_S1_val <- preProcess(taxoS1_adj_val, method = "range")

taxoS_adj_norm_train <- predict(preProcValues_S_train, taxoS_adj_train)
taxoS1_adj_norm_train <- predict(preProcValues_S1_train, taxoS1_adj_train)
taxoS_adj_norm_val <- predict(preProcValues_S_val, taxoS_adj_val)
taxoS1_adj_norm_val <- predict(preProcValues_S1_val, taxoS1_adj_val)
```

Obtenim la trasposada de les matrius perquè puguin utilitzar-se en ML/DL.

```{r}
taxoS_train <- transpose(taxoS_adj_norm_train)
taxoS1_train <- transpose(taxoS1_adj_norm_train)
taxoS_val <- transpose(taxoS_adj_norm_val)
taxoS1_val <- transpose(taxoS1_adj_norm_val)
taxoS_test <- transpose(taxoS_adj_norm_test)
taxoS1_test <- transpose(taxoS1_adj_norm_test)

colnames(taxoS_train) <- rownames(taxoS_adj_norm_train)
taxoS_train$sample <- colnames(taxoS_adj_norm_train)
taxoS_train <- taxoS_train %>% select("sample", everything())
colnames(taxoS1_train) <- rownames(taxoS1_adj_norm_train)
taxoS1_train$sample <- colnames(taxoS1_adj_norm_train)
taxoS1_train <- taxoS1_train %>% select("sample", everything())

colnames(taxoS_val) <- rownames(taxoS_adj_norm_val)
taxoS_val$sample <- colnames(taxoS_adj_norm_val)
taxoS_val <- taxoS_val %>% select("sample", everything())
colnames(taxoS1_val) <- rownames(taxoS1_adj_norm_val)
taxoS1_val$sample <- colnames(taxoS1_adj_norm_val)
taxoS1_val <- taxoS1_val %>% select("sample", everything())

colnames(taxoS_test) <- rownames(taxoS_adj_norm_test)
taxoS_test$sample <- colnames(taxoS_adj_norm_test)
taxoS_test <- taxoS_test %>% select("sample", everything())
colnames(taxoS1_test) <- rownames(taxoS1_adj_norm_test)
taxoS1_test$sample <- colnames(taxoS1_adj_norm_test)
taxoS1_test <- taxoS1_test %>% select("sample", everything())
```

Guardem arxius al disc.

```{r}
write.csv(taxoS_train,"output/taxoS_train.csv", row.names = FALSE)
write.csv(taxoS_val,"output/taxoS_val.csv", row.names = FALSE)
write.csv(taxoS_test,"output/taxoS_test.csv", row.names = FALSE)
write.csv(taxoS1_train,"output/taxoS1_train.csv", row.names = FALSE)
write.csv(taxoS1_val,"output/taxoS1_val.csv", row.names = FALSE)
write.csv(taxoS1_test,"output/taxoS1_test.csv", row.names = FALSE)
```
