==> Avaluació_MIPredictor/04_Avalua_models_alternatius.ipynb <==
#!pip install tensorflow
#!pip install sdv

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import tensorflow as tf
import sdv
import joblib
from joblib import dump, load

from sdv.tabular import GaussianCopula, CTGAN, CopulaGAN, TVAE
from sdv.sampling import Condition
from sdv.evaluation import evaluate

from numpy import random
from matplotlib.pyplot import figure
from sklearn.preprocessing import MinMaxScaler, RobustScaler
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV, KFold
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn import metrics
from tensorflow import keras
from tensorflow.keras import layers, losses
from tensorflow.keras.models import Model, Sequential, save_model, load_model
from tensorflow.keras.layers import Dense, Input, Conv1D, Activation, Reshape, Flatten, Dropout, MaxPooling1D
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint
from keras.wrappers.scikit_learn import KerasClassifier
from sklearn.metrics import confusion_matrix, accuracy_score, roc_auc_score, roc_curve

def which_taxo(file):

    if (file.find("a_") != -1):
        taxo = pd.read_csv("test/taxoS_test.csv")
        nom = "CRC S"
    elif (file.find("b_") != -1):
        taxo = pd.read_csv("test/taxoS1_test.csv")
        nom = "CRC S1"
    elif (file.find("k_") != -1):
        taxo = pd.read_csv("test/taxoS_test.csv")
        nom = "CRC + Regió S"
    elif (file.find("l_") != -1):
        taxo = pd.read_csv("test/taxoS1_test.csv")
        nom = "CRC + Regió S1"
    elif (file.find("m_") != -1):
        taxo = pd.read_csv("test/taxoS_test.csv")
        nom = "CRC + Seq_tool S"
    elif (file.find("n_") != -1):
        taxo = pd.read_csv("test/taxoS1_test.csv")
        nom = "CRC + Seq_tool S1"
    elif (file.find("o_") != -1):
        taxo = pd.read_csv("test/taxoS_test.csv")
        nom = "CRC + Regió + Seq_tool S"
    elif (file.find("p_") != -1):
        taxo = pd.read_csv("test/taxoS1_test.csv")
        nom = "CRC + Regió + Seq_tool S1"
    return taxo, nom

def which_meta_extra(file):
    if (file.find("m_") != -1) or (file.find("n_") != -1):
        meta_extra = ["seq_tool"]
    
    elif (file.find("o_") != -1) or (file.find("p_") != -1):
        meta_extra = ["region","seq_tool"]
        
    elif (file.find("k_") != -1) or (file.find("l_") != -1):
        meta_extra = ["region"]
        
    else:
        meta_extra = []
    return meta_extra

def filt_y_rows(taxo,meta):
    condition = []
    for i in range(len(taxo)):
        condition.append(pd.DataFrame(meta["condition"][meta["sample"] == 
                                                        taxo["sample"][i]]).iloc[0,0])
    return condition

def add_metas(taxo,meta,meta_extra=[]):
    for i in meta_extra:
        meta_extra_i = []
        for j in range(len(taxo)):
            meta_extra_i.append(pd.DataFrame(meta[i][meta["sample"] ==
                                          taxo["sample"][j]]).iloc[0,0])
        taxo[i] = meta_extra_i
    return taxo

def drop_nas(taxo,meta):
    taxo["condition"] = meta
    taxo = taxo.dropna()
    meta = taxo["condition"]
    taxo = taxo.iloc[:,1:-1]
    return taxo, meta

def get_roc_curves(y_test, y_prob, pos_label):
    fper1, tper1, tresholds1 = roc_curve(y_test, y_prob, pos_label=pos_label)
    fig,base = plt.subplots()
    base.set_title("Corba ROC" + " " + nomi)
    base.set_xlabel("Especificitat")
    base.set_ylabel("Sensibilitat")
    base.plot(fper1, tper1)
    base.plot([0,1],[0,1],'-')
    fig.savefig("metrics/" + file + '_roc_curve.png')
    plt.close()

def get_test_metrics(file, model, x_test, y_test, labels):
    y_pred = model.predict(x_test)
    if (file.find(".h5") == -1):
        y_prob = model.predict_proba(x_test)[:,1]
    else:
        y_prob = y_pred
    y_pred = pd.DataFrame(np.around(y_pred,0).astype(int))
    y_pred.replace([0,1], labels, inplace=True)
    tn, fp, fn, tp = confusion_matrix(y_test, y_pred, labels = labels).ravel()
    acc = accuracy_score(y_test, y_pred)
    sens = tp / (tp+fn)
    espe = tn / (tn+fp)
    AUC = 1 - roc_auc_score(y_test, y_prob)
    get_roc_curves(y_test, y_prob, labels[1])
    return acc, sens, espe, AUC
    
def encode_data(x_test, meta_extra, autoencoder):
        if len(meta_extra) == 0:
            x_test_no_meta = x_test
        else:
            x_test_no_meta = x_test.iloc[:,0:(-len(meta_extra))]
        x_test_no_meta = x_test_no_meta.to_numpy()
        x_test_no_meta = np.expand_dims(x_test_no_meta, axis=2)
        x_test_no_meta = x_test_no_meta.astype('float32')
        encoded_data = autoencoder.encoder(x_test_no_meta).numpy()
        encoded_data = pd.DataFrame(encoded_data)
        
        if len(meta_extra) == 2:
            encoded_data["region"] = x_test.reset_index(drop=True).iloc[:,-1]
            encoded_data["seq_tool"] = x_test.reset_index(drop=True).iloc[:,-2]
        elif len(meta_extra) == 1:
            if meta_extra == ["region"]:
                encoded_data["region"] = x_test.reset_index(drop=True).iloc[:,-1]
            else:
                encoded_data["seq_tool"] = x_test.reset_index(drop=True).iloc[:,-1]

        scaler = MinMaxScaler()
        scaler.fit(encoded_data)
        encoded_data_scaled = pd.DataFrame(scaler.transform(encoded_data))
        return encoded_data_scaled

taxo = pd.read_csv("test/taxoS1_test.csv")
len(taxo.columns)

DirExists = os.path.exists("metrics")
if not DirExists:
    os.makedirs("metrics")
nom = []
nom_model = []
N = []
acc = []
sens = []
espe = []
AUC = []
for file in os.listdir("models/altres_models/"):
    if (file.find("autoencoders") == -1 & file.find("synthetizers") == -1 & file.find("params") == -1 & file.find("checkpoints")==-1):
        print(file)
        x_test, nomi = which_taxo(file)
        y = pd.read_csv("metadades_full.csv")
        meta_extra = which_meta_extra(file)
        y_test = filt_y_rows(x_test, y)
        x_test = add_metas(x_test, y, meta_extra)
        x_test, y_test = drop_nas(x_test, y_test)
        if (file.find("rawdata") == -1):
            autoencoder_filename = "models/altres_models/autoencoders/" + file[0] + "_autoencoder"
            autoencoder = tf.saved_model.load(autoencoder_filename)
            x_test = encode_data(x_test, meta_extra, autoencoder)
            if (file.find("encoded") == -1):
                if (file.find("synthetic") != -1):
                    nomi2 = "(Synthetic)"
                elif (file.find("synreal") != -1):
                    nomi2 = "(Syn+Real)"
            else:
                nomi2 = "(Encoded)"
        else:
            nomi2 = "(Rawdata)"
        nomi = nomi + " " + nomi2
        if (file.find(".joblib") == -1):
            model = load_model("models/altres_models/" + file)
        else:
            model = joblib.load("models/altres_models/" + file)
            
        N.append(len(x_test))
        nom.append(nomi)
        nom_model.append(str(model)[0:10])
        acci, sensi, espei, AUCi =  get_test_metrics(file, model, x_test, y_test, ["Control","CRC"])
        acc.append(acci)
        sens.append(sensi)
        espe.append(espei)
        AUC.append(AUCi)

results = pd.DataFrame()
results["Nom"] = nom
results["n (test)"] = N
results["Model"] = nom_model
results["Exactitud"] = acc
results["Sensibilitat"] = sens
results["Especificitat"] = espe
results["AUC"] = AUC

results.to_csv("metrics/metriques_models_alternatius.csv", index=False)

print("metrics done")

results[results["Nom"].str.contains('Rawdata')]

results[results["Nom"].str.contains('Encoded')]

results[results["Nom"].str.contains('Synthetic')]

results[results["Nom"].str.contains('Real')]
